2020/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cymruwhois==1.6
dnspython==2.0.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1
